+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 * [EACQAPW-5532] TDSCertsManager Tests
 * [EACQAPW-5551] Move to Static linking
 * [EACQAPW-5556] CardPaymentProcess Tests
+* [EACQAPW-5562] ExternalAPIClient Tests
+* [EACQAPW-5581] CardDataFormatter Tests
+* [EACQAPW-5583] URLDataLoaderTests Tests
 
 ### Fixed